backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic-settings # Adicionado para resolver o erro

# Novas dependências adicionadas
apscheduler
geopy
pyppeteer # Substituindo weasyprint
jinja2

# Ferramentas de Desenvolvimento
ruff